+++ docs/src/examples/components/Reaction/Rtl/index.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react'
+
+import ComponentExample from 'docs/src/components/ComponentDoc/ComponentExample'
+import NonPublicSection from 'docs/src/components/ComponentDoc/NonPublicSection'
+
+const Rtl = () => (
+  <NonPublicSection title="Rtl">
+    <ComponentExample examplePath="components/Reaction/Rtl/ReactionExample.rtl" />
+  </NonPublicSection>
+)
+
+export default Rtl