+++ packages/react/test/specs/components/Dropdown/Dropdown-test.tsx
@@ -183,4 +183,15 @@ describe('Dropdown', () => {
       )
     })
   })
+
+  describe('getA11ySelectionMessage', () => {
+    it('creates message container element', () => {
+      mountWithProvider(<Dropdown options={[]} getA11ySelectionMessage={{}} />)
+      expect(
+        document.querySelector(
+          `[role="status"][aria-live="polite"][aria-relevant="additions text"]`,
+        ),
+      ).toBeTruthy()
+    })
+  })
 })